requirements.txt
openai-whisper==20240930
groq==0.15.0
python-dotenv==1.0.1
sounddevice==0.5.1
numpy<2.0.0
requests>=2.32.0
streamlit>=1.41.1
llvmlite<0.40.0
numba<0.57.0
